BiblioTech/requirements.txt
django
djangorestframework
python-dotenv
markdown
django-filter
psycopg
psycopg2
django-bootstrap-v5
django-crispy-forms
crispy-bootstrap5